+++ CHANGELOG.md
@@ -2,7 +2,14 @@
 
 ## [Unrealeased]
 
+### Added
+
 - [TD-4463] add notifications for `rule_created` and `implementation_created`
+
+### Fixed
+
+- [TD-4427] Fixed a performance issue when calling /api/executions/search
+  by filtering by sources on the database, avoiding a potential time out.
 
 ## [4.37.1] 2022-02-10